Assignment_3/requirements.txt
pymongo>=4.6
python-dotenv>=1.0
requests>=2.31